row: add compact_and_expire() variant which collects purged tombstones
The new variant is called
`compact_expire_and_collect_purged_tombstones()`. The difference
compared to the existing `compact_and_expire()` variant is that this one
collects all the purged cells and tombstones into a new row object and
returns it to the caller. This can be used to retain these dead cells and
tombstones temporarily until some long process (like compaction) finishes
safely, preventing data resurrection in the case of an interruption.