size_tiered_backlog_tracker: Add backlog validator
This validator helped debugging issues like #6021 and #6227 where
there was bugs in the tracker implementation, which results in
the backlog calculated by the tracker not matching the expected
backlog which would be a perfect calculation of it respecting
all the properties of the formula.

This validator works by calculating the backlog from scratch,
not using any accumulators like the tracker implementation does,
comparing the result to the one produced by the tracker, and
reporting any significant mismatch.

Can be enabled via flag COMPACTION_BACKLOG_TRACKER_DEBUG

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>