compaction/twcs: Make task estimation take into account the size-tiered behavior
The task estimation was not taking into account that TWCS does size-tiered
on the current window, and it only added 1 to the estimation when there
could be more tasks than that depending on the amount of SSTables in
all the size tiers. Let's also adjust task estimation to take into
account the new behavior where past windows can now also be compacted
with size-tiered.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>