table: Introduce off-strategy compaction on repair set
Off-strategy compaction is about incrementally reshaping the sstables
in repair set, using our existing reshape mechanism, until the set
is ready for integration into the main sstable set. It's done in
maintanance mode. We can do it this way because data in repair
sstable set is potentially disjoint and so read amplification is
avoided by using partitioned_sstable_set.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>