compaction_strategy: Introduce split_during_flush
This option is inherited from C* and it's used to determine if segregation
should be performed on memtable flush. The intention is to start with it
disabled by default, allow users to enable it experimentally, and once
we're confident about it, then enable it by default.

usage:

create table ... with compaction = {
	'class': 'SizeTieredCompactionStrategy',
	'split_during_flush': 'true|false'
};

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>