table: Introduce maintenance sstable set
This new set will hold the sstables created by maintenance operations
like bootstrap and replace.
Given that these new sstables don't overlap, partitioned_sstable_set
is used to avoid read amplification.
Also, this new set will not be accounted by backlog tracker as we're
able to store the new sstables efficiently and can compact them in
maintenance mode, so no rush.

Subsequent work will do more integration of this new set like including
it in the read path, accouting it for statistics, and so on.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>