compaction: Reduce backlog of compacting SSTable properly
It was observed that as compaction progresses the backlog of compacting SSTable
is being reduced very slowly, which causes shares to be higher than needed, and
consequently compaction acts much more aggressively than it has to.

https://user-images.githubusercontent.com/1409139/120237819-93dfc080-c232-11eb-9042-68114e285ea0.png

The graph above shows the amount of backlog that is reduced from a SSTable
being compacted. The red line denotes the total backlog of the SSTable, before
it's selected for compaction. The expectation is that the more a SSTable is
compacted the more backlog will be reduced from it. However, in the current
implementation, it can be seen that the backlog to be reduced, from the SSTable
being compacted, starts being inversely proportional to the amount of data
already compacted.

Turns out that this problem happens because the implementation of backlog
formula becomes incorrect when the SSTable is being compacted.

Backlog for a sstable is defined as:
    Bi = Ei * log (T / Ei)

    where Ei = Si - Ci (bytes left to be compacted)
        and Si = size of SStable
        and Ci = total bytes compacted
        and T = total size of table

The formula above can also be rewritten as follows:
    Bi = Ei * log (T) - Ei * log (Ei)

the second term `Ei * log (Ei)` can be rewritten as:
    = (Si - Ci) * log (Ei)
    = Si * log (Ei) - Ci * log (Ei)

However, digging backlog implementation, turns out that we're incorrectly
implementing that second term as:
    = Si * log (Si) - Ci * log (Ei)

Given that Si > Ei, for a SSTable being compacted, the backlog will be higher
than it should.

the following table shows how the backlog of a SSTable being compacted behaves:
https://gist.github.com/raphaelsc/42e14be0d7d4ed264e538c2d217c8f95

It can be clearly seen that backlog behaves weirdly as described above. With my
fix, backlog of compacting sstable will behave exactly as we expect it to.

When a SSTable is added to the backlog tracker, it contributes with a backlog
of `Si * log(Si)` (see size_tiered_backlog_tracker::add_sstable()).

Given that sstable's backlog formula is Bi=(Si - Ci) * log(Ei),
    where Ei = Si - Ci

We need to subtract (Si * log(Si / (Si - Ci))) from the initial `Si * log(Si)`,
    to produce (Si * log(Si - Ci)) = Si * log(Ei).

Example:
1) SSTable A is created, contributes with a backlog of Si * log(Si)
    Given that Ei = Si and Ci = 0, Si * log(Si) = (Si - Ci) * log (Ei)
2) SSTable A is now being compacted, so Ei != Si, Ci > 0, so
    1) `(Si * log(Si / (Si - Ci)))` is subtracted from the initial
    `Si * log(Si)`, so we end up with Si * log(Ei).
    2) `Ci * log(Ei)` is also subtracted from the initial `Si * log(Si)`

    By doing both subtractions above, we achieve the target formula for the
    second term, see:
        = Si * log(Si) - (Si * log(Si / (Si - Ci))) - Ci * log(Si - Ci)
        = Si * log(Si - Ci) - Ci * log(Si - Ci)
        = Si * log(Ei) - Ci * log(Ei)
        = (Si - Ci) * log (Ei)

The fix is a bit complex, but it was the best one I could come up up which will
be easy to backport. The easier one involves removing SSTables from the backlog
tracker in the beginning of compaction, so backlog of compacting SSTable would
be easy to calculate, simply (Si - Ci) * log(Ei), but that requires touching
an area in compaction that can easily break things, so wouldn't be safe to
backport it at all.

Generated the following graph showing that backlog behavior is now fixed:
https://user-images.githubusercontent.com/1409139/120254044-4d548b00-c25f-11eb-8da3-8c204b7d00b5.png

Those sudden decreases in the current (broken) version is due to a bug (also
fixed in this patch) in the calculation at the exact moment input sstable is
fully compacted, so log4(0) is called which produces -inf. Backlog doesn't go
negative because 0 is the minimum allowed.

Fixes #8768.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>