compaction: Don't leak backlog of input sstable when compaction strategy is changed
When compaction strategy is changed, we decided to ignore the ongoing
compactions when calculating backlog as the new compaction strategy
has different compaction requirements.

We implement that by:
1) not registering ongoing compactions in the new backlog tracker.
2) asking ongoing compaction to stop tracking input sstables

turns out that #2 is incorrect. It will only cause input sstables to
be leaked.

That's because transfer_ongoing_charges() will already exclude ongoing
compactions from the new tracker.

For #2, stop_tracking_ongoing_compactions() will set the tracking flag
of ongoing compaction to false, which will cause
compaction_read_monitor to skip removal of input sstable from the
tracker. So given that input sstable is removed from the sstable set
when compaction ends, but never removed from the tracker, compaction
shares will be higher than needed.

To fix that, we'll only stick to #1, ongoing compactions aren't
registered, but ongoing compactions should still remove their input
sstables from the tracker in the end.

Refs #9157.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>