table: Make set_compaction_strategy() exception safe
to make it exception safe, new backlog tracker is only registered
when we're done building new sstable set and transferring charges
to the new tracker.
transfer_ongoing_charges() will no longer update the old tracker,
which is fine because the old tracker will be deregistered as
soon as new one replaces it.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>