compaction_strategy: Allow strategies to work with sstable runs
This new interface will allow compaction strategies to work with either
runs or sstables. Existing interface is kept for compatibility reasons,
but intended to be deprecated and later removed.

The overhead of converting runs to sstables isn't a problem because
compaction manager will no longer copy non-compacting sstables into a
candidate list, but will forward sstable runs, retrieved from sstable_set,
directly into the compaction strategy. So the copy is only being moved
from one place to another.

This new interface will benefit ICS and if we end up changing LCS
to work with runs, which is a result of decoupling physical level from
logical one.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>