compaction: place compaction result data into its own structure
compaction_info is used to track compaction activity, but it's also
being used to place output data returned by compact_sstables().
Let's move output data into a structure of its own, such that each
structure is not mixed with data for different purposes.
This is also a step towarding fixing a layer violation, where
compaction is aware of compaction manager.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>