db: add a distributed table to store shared sstables ownership info
Introduce a new fully replicated `shared_sstables` table
in the `system_distributed_everywhere` keyspace (RF=ALL).

It will serve as a distributed reference counting mechanism
to track uses of shared sstables in an external object storage.

The intent is that when no one is holding a reference to an
sstable, it can safely be deleted from an object storage.

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>