db: add methods to add and remove owners for shared sstables
Use LWT to maintain consistent state of
`system_distributed_everywhere.shared_sstables` table.

When removing an owner for an sstable, it is checked
whether there are any references left. If there are none,
the `remove_shared_sstable_owner` method will return
`false` to indicate that the caller should issue a command
to remove the sstable from an external object storage.

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>