table: Clear backlog tracker when stopping table
Backlog tracker is managed by compaction strategy, and we'd like to
have it cleared in table::stop(), to make sure that all state is
cleared too. For example, a reference to shared sstable could
prevent sstable manager from being stopped as it relies on all
sstables managed by it being closed first, and that has to happen
within table::stop().

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>