compaction: Redefine size-tiered compaction backlog to fix unnecessary aggressiveness
Today, compaction can act much more aggressive than it really has to, because
the strategy and its definition of backlog are completely decoupled.

The backlog definition for size-tiered, which is inherited by all
strategies (LCS' L0, TWCS' windows), is built on the assumption that the
world must reach the state of zero amplification. But that's unrealistic
and goes against the intent amplification defined by the compaction
strategy. For example, size tiered is a write oriented strategy which
allows for extra space amplification for compaction to keep up with the
relatively high write rate.

It can be seen today, in many deployments, that compaction shares is
either close to 1000, or even stuck at 1000, even though there's nothing
to be done. When there's a single sstable per tier, for example.
This means that whenever a new compaction job kicks in, it will act much
more aggressive because of the high shares, caused by false backlog of
the existing tables.

This problem can be fixed, as proposed in my doc "Fixing compaction
aggressiveness due to suboptimal definition of zero backlog by controller",
by removing backlog of tiers that don't have to be compacted now, like a tier
that has a single file. Low-efficiency compaction, like compacting 2 files only
or cross-tier, only happens when system is under little load and can proceed at
a slower pace. Once efficient jobs show up, ongoing compactions, even if
inneficient, will get more shares (as efficient jobs add to the backlog) so
compaction won't fall behind.
With this approach, throughput and latency is improved as cpu time is no
longer stolen (unnecessarily) from the foreground requests.

Fixes #4588.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>