sstable_set: keep sstable_set_impl_ptr in incremental selectors impl
Make the incremental selector implementations self-sufficient
by keeping a shared ptr of the sstable_set_impl so they can
select from a read-only snapshot of themselves.

With that, there is no longer a need to
keep the lw_shared_ptr<sstable_set> in the selectors.

The reason we keep the sstable_set_impl_ptr in the
incremental_selector_impl and not in sstable_set::incremental_selector
is for compound_sstable_set optimization for a single
non-empty set (single compaction_group with empty maintenance_set)
when compound_sstable_set::make_incremental_selector needs to
return an incremental_selector_impl - so it needs to be
self sufficient and keep a shared_ptr of the underlying impl.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>