sstable_set: allow updating in-place
Make the sstable_set equivalent to a smart pointer
of the sstable_set_impl.

With that, we no longer need to keep a lw_shared_ptr<sstable_set>
for an immutable copy of it, but rather we just need to
copy the sstable_set, and get a shared pointer to its sstable_set_impl.

To modify it, we first get a clone(), modify the teamporary copy
and finally update the original sstable_set in place
by move-assigning the modified sstable_set into the original one.

A following patch will change compound_sstable_set to
point to the underlying sstable sets so that can be updated in place
without having to rebuild the compound sstable set.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>