sstable_set: compound_sstable_set: point to underlying sstable_set:s
So that the compound set doesn't have to be rebuilt
when any of the underlying sstable_set:s are updated,
now that the update is in place.

With that, get rid of table::refresh_compound_sstable_set().

table::sstables_as_snapshot_source now needs to clone
the _sstables set since it is updated in place in update_cache.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>